Pushing sleuthkit SLOT back to 0
Even though the libraries and several of the binaries in sleuthkit-3 change and
could peacefully coexist with earlier branches, there is still sufficient
overlap to justify leaving it as a single SLOT.  One day down the road if/when
Carrier takes the "one binary, many args" approach like git et. al. that several
have suggested, it may be more possible.