Bump dev-util/radare to 1.5.2
This is wholly bad form, I'm just bumping it to scratch an itch and see if it
gets rid of the hard-coded python-2.5 and python-2.6 directories (it doesn't).
The only reason I'm keeping it is because I put time into it and added the 'ewf'
USE flag so it appropriately links against libewf when desired.  That said, this
thing is a mess - QA warnings all over the place, static python plugins,
non-orthogonal configuration options (hard-coded their own --with-X logic
instead of following the autotools standard).  I have no desire to go through
and pick this apart completely.