mass removal of libyal deps
Although libyal-r1.eclass is proving to be useful, the same issues that popped
up last time we tried this are popping up again.

In short, libyal has multiple tiers of libraries.  Each time a release is cut
for, say, a first-tier library, that release carries snapshots of its dependency
libraries along with it.  It can (and enthusiastically will) use a system-wide
shared copy, but evidently libyal development is still moving quickly enough
that different top-tier libraries (like libewf and libbde) carry subtly
different copies and are developed _to that copy_.

Having a global libcsystem and libfdata are a nice idea and will likely one day
work, but right now libc* and libf* are out as direct dependencies.  Bigger
binaries, fewer problems.